Created folder for the pages
